+++ src/egon/data/airflow/dags/pipeline.py
@@ -11,6 +11,7 @@ from egon.data.processing.zensus_vg250 import (
 )
 import airflow
 import egon.data.importing.demandregio as import_dr
+import egon.data.importing.demandregio.install_disaggregator as install_dr
 import egon.data.importing.etrago as etrago
 import egon.data.importing.heat_demand_data as import_hd
 import egon.data.importing.mastr as mastr
@@ -182,10 +183,20 @@ with airflow.DAG(
 
     scenario_input_tables >> demandregio_tables
 
+
+    demandregio_installation = PythonOperator(
+        task_id="demandregio-installation",
+        python_callable=install_dr.clone_and_install,
+    )
+
+    setup >> demandregio_installation
+
     demandregio_society = PythonOperator(
         task_id="demandregio-society",
         python_callable=import_dr.insert_society_data,
     )
+
+    demandregio_installation >> demandregio_society
     vg250_clean_and_prepare >> demandregio_society
     demandregio_tables >> demandregio_society
     scenario_input_import >> demandregio_society
@@ -194,6 +205,8 @@ with airflow.DAG(
         task_id="demandregio-household-demands",
         python_callable=import_dr.insert_household_demand,
     )
+
+    demandregio_installation >> demandregio_demand_households
     vg250_clean_and_prepare >> demandregio_demand_households
     demandregio_tables >> demandregio_demand_households
     scenario_input_import >> demandregio_demand_households
@@ -202,6 +215,8 @@ with airflow.DAG(
         task_id="demandregio-cts-industry-demands",
         python_callable=import_dr.insert_cts_ind_demands,
     )
+
+    demandregio_installation >> demandregio_demand_cts_ind
     vg250_clean_and_prepare >> demandregio_demand_cts_ind
     demandregio_tables >> demandregio_demand_cts_ind
     scenario_input_import >> demandregio_demand_cts_ind